# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/sessions/{sessionId}/templates": {
      "post": {
        "operationId": "TemplateController_create",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTemplateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Template created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateResponseDto"
                }
              }
            }
          }
        },
        "summary": "Create a message template for the session (text or with a single media attachment)",
        "tags": [
          "templates"
        ]
      },
      "get": {
        "operationId": "TemplateController_findBySession",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of templates",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TemplateResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "List all templates for a session",
        "tags": [
          "templates"
        ]
      }
    },
    "/api/sessions/{sessionId}/templates/{id}": {
      "get": {
        "operationId": "TemplateController_findOne",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Template ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Template details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Template not found"
          }
        },
        "summary": "Get a template by ID",
        "tags": [
          "templates"
        ]
      },
      "put": {
        "operationId": "TemplateController_update",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Template ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTemplateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Template updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Template not found"
          }
        },
        "summary": "Update a template (including attaching, replacing, or removing its media)",
        "tags": [
          "templates"
        ]
      },
      "delete": {
        "operationId": "TemplateController_delete",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Template ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Template deleted"
          },
          "404": {
            "description": "Template not found"
          }
        },
        "summary": "Delete a template",
        "tags": [
          "templates"
        ]
      }
    },
    "/api/sessions/{sessionId}/templates/{id}/media": {
      "get": {
        "operationId": "TemplateController_getMedia",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Template ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Raw media bytes"
          },
          "404": {
            "description": "Template not found or has no media"
          }
        },
        "summary": "Stream a template's media attachment (for dashboard preview)",
        "tags": [
          "templates"
        ]
      }
    },
...
  },
  "info": {
    "title": "OpenWA API",
    "description": "Open Source WhatsApp API Gateway - Free, Self-Hosted HTTP API",
    "version": "0.10.10",
    "contact": {
      "name": "OpenWA",
      "url": "https://github.com/rmyndharis/OpenWA",
      "email": "yudhi@rmyndharis.com"
    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "X-API-Key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "schemas": {
...
      "CreateTemplateDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Unique template name within the session",
            "example": "order-confirmation",
            "maxLength": 100
          },
          "body": {
            "type": "string",
            "description": "Template body with {{variable}} placeholders. Required for a text template; optional when a media attachment is provided (the body then renders as the media caption, and may be empty). The \"body required unless media is attached\" rule is enforced by the service.",
            "example": "Hi {{customer}}, your order {{orderId}} has shipped.",
            "maxLength": 4096
          },
          "header": {
            "type": "string",
            "description": "Optional header text, prepended to the rendered body",
            "example": "OpenWA Store",
            "maxLength": 1024
          },
          "footer": {
            "type": "string",
            "description": "Optional footer text, appended to the rendered body",
            "example": "Reply STOP to unsubscribe.",
            "maxLength": 1024
          },
          "mediaType": {
            "type": "string",
            "description": "Attach a single media file to the template. When set, the template is sent as this media type with the rendered text as its caption. Requires mediaBase64 and mimetype.",
            "enum": [
              "image",
              "video",
              "document",
              "audio"
            ],
            "example": "image"
          },
          "mediaBase64": {
            "type": "string",
            "description": "Base64-encoded media bytes (data: URI prefix accepted). Required when mediaType is set."
          },
          "mimetype": {
            "type": "string",
            "description": "MIME type of the attached media. Required when mediaType is set.",
            "example": "image/jpeg"
          },
          "filename": {
            "type": "string",
            "description": "Optional filename for the attached media",
            "example": "promo.jpg",
            "maxLength": 255
          }
        },
        "required": [
          "name"
        ]
      },
      "TemplateResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "sessionId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "body": {
            "type": "string",
            "nullable": true
          },
          "header": {
            "type": "string",
            "nullable": true
          },
          "footer": {
            "type": "string",
            "nullable": true
          },
          "hasMedia": {
            "type": "boolean",
            "description": "True when the template carries a media attachment."
          },
          "mediaType": {
            "type": "string",
            "enum": [
              "image",
              "video",
              "document",
              "audio"
            ],
            "nullable": true
          },
          "mimetype": {
            "type": "string",
            "nullable": true
          },
          "filename": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "sessionId",
          "name",
          "hasMedia",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateTemplateDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Template name",
            "maxLength": 100
          },
          "body": {
            "type": "string",
            "description": "Template body with {{variable}} placeholders",
            "maxLength": 4096
          },
          "header": {
            "type": "string",
            "description": "Optional header text",
            "maxLength": 1024
          },
          "footer": {
            "type": "string",
            "description": "Optional footer text",
            "maxLength": 1024
          },
          "mediaType": {
            "type": "string",
            "description": "Replace the media attachment. Pass mediaType=null (with removeMedia) to clear it, or a new mediaType + mediaBase64 + mimetype to swap it. Omit these fields to leave the media unchanged.",
            "enum": [
              "image",
              "video",
              "document",
              "audio"
            ],
            "example": "video"
          },
          "mediaBase64": {
            "type": "string",
            "description": "Base64-encoded media bytes. Required when replacing media (mediaType set)."
          },
          "mimetype": {
            "type": "string",
            "description": "MIME type of the attached media. Required when replacing media.",
            "example": "video/mp4"
          },
          "filename": {
            "type": "string",
            "description": "Optional filename for the attached media",
            "maxLength": 255
          },
          "removeMedia": {
            "type": "boolean",
            "description": "Set true to remove the existing media attachment (turning the template back into text-only).",
            "example": false
          }
        }
      },
...
    }
  },
  "security": [
    {
      "X-API-Key": []
    }
  ]
}